Removed kubernetes related files
