+++ e2e/provider.go
@@ -4,7 +4,6 @@ import (
 	"github.com/e2eterraformprovider/terraform-provider-e2e/client"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/blockstorage"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/image"
-	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/kubernetes"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/loadbalancer"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/node"
 	"github.com/e2eterraformprovider/terraform-provider-e2e/e2e/objectstore"
@@ -50,7 +49,6 @@ func Provider() *schema.Provider {
 			"e2e_sfs":          sfs.ResourceSfs(),
 			"e2e_objectstore":  objectstore.ResourceObjectStore(),
 			"e2e_ssh_key":      ssh_key.ResourceSshKey(),
-			"e2e_kubernetes":   kubernetes.ResourceKubernetesService(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"e2e_node":            node.DataSourceNode(),
@@ -63,7 +61,6 @@ func Provider() *schema.Provider {
 			"e2e_reserve_ips":     reserve_ip.DataSourceReserveIps(),
 			"e2e_sfss":            sfs.DataSourceSfs(),
 			"e2e_objectstores":    objectstore.DataSourceObjectStores(),
-			"e2e_kubernetes":      kubernetes.DataSourceKubernetesService(),
 		},
 		ConfigureFunc: providerConfigure,
 	}
